requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
aiohttp==3.7.4.post0
    # via cdip-connector
async-timeout==3.0.1
    # via aiohttp
attrs==21.2.0
    # via aiohttp
cachetools==4.2.2
    # via google-auth
https://github.com/PADAS/smart-integrate-sdk/releases/download/v0.4-alpha/cdip_connector-0.4.0-py3-none-any.whl
    # via -r requirements.in
certifi==2021.5.30
    # via requests
chardet==4.0.0
    # via
    #   aiohttp
    #   requests
click==7.1.2
    # via uvicorn
https://github.com/PADAS/das-clients/releases/download/v1.0.9/dasclient-1.0.9-py3-none-any.whl
    # via -r requirements.in
environs==9.3.2
    # via
    #   -r requirements.in
    #   cdip-connector
fastapi==0.65.1
    # via -r requirements.in
google-api-core[grpc]==1.30.0
    # via google-cloud-pubsub
google-auth==1.30.2
    # via google-api-core
google-cloud-pubsub==2.5.0
    # via -r requirements.in
googleapis-common-protos[grpc]==1.53.0
    # via
    #   google-api-core
    #   grpc-google-iam-v1
grpc-google-iam-v1==0.12.3
    # via google-cloud-pubsub
grpcio==1.38.0
    # via
    #   google-api-core
    #   googleapis-common-protos
    #   grpc-google-iam-v1
h11==0.12.0
    # via uvicorn
idna==2.10
    # via
    #   requests
    #   yarl
libcst==0.3.19
    # via google-cloud-pubsub
marshmallow==3.12.1
    # via environs
multidict==5.1.0
    # via
    #   aiohttp
    #   yarl
mypy-extensions==0.4.3
    # via typing-inspect
packaging==20.9
    # via google-api-core
proto-plus==1.18.1
    # via google-cloud-pubsub
protobuf==3.17.3
    # via
    #   google-api-core
    #   googleapis-common-protos
    #   proto-plus
pyasn1-modules==0.2.8
    # via google-auth
pyasn1==0.4.8
    # via
    #   pyasn1-modules
    #   rsa
pydantic==1.8.2
    # via
    #   -r requirements.in
    #   cdip-connector
    #   fastapi
pyparsing==2.4.7
    # via packaging
python-dateutil==2.8.1
    # via cdip-connector
python-dotenv==0.17.1
    # via environs
python-json-logger==2.0.1
    # via cdip-connector
pytz==2021.1
    # via google-api-core
pyyaml==5.4.1
    # via libcst
redis==3.5.3
    # via walrus
requests==2.25.1
    # via
    #   -r requirements.in
    #   cdip-connector
    #   dasclient
    #   google-api-core
rsa==4.7.2
    # via google-auth
six==1.16.0
    # via
    #   google-api-core
    #   google-auth
    #   grpcio
    #   protobuf
    #   python-dateutil
starlette==0.14.2
    # via fastapi
statsd==3.3.0
    # via cdip-connector
typing-extensions==3.10.0.0
    # via
    #   aiohttp
    #   libcst
    #   pydantic
    #   typing-inspect
typing-inspect==0.7.1
    # via libcst
urllib3==1.26.5
    # via requests
uvicorn==0.13.4
    # via -r requirements.in
walrus==0.8.2
    # via -r requirements.in
yarl==1.6.3
    # via aiohttp

# The following packages are considered to be unsafe in a requirements file:
# setuptools
